CANS-343. Refactor the domain progression bar. Use filled/total presentation instead of percentage
