+++ app/javascript/Application/components/Assessment/AssessmentHelper.test.js
@@ -3,7 +3,6 @@ import {
   resetConfidentialByDefaultItems,
   shouldDomainBeRendered,
   shouldItemBeRendered,
-  calculateDomainProgress,
 } from './AssessmentHelper';
 import { clone } from '../../util/common';
 
@@ -204,107 +203,6 @@ describe('AssessmentHelper', () => {
 
       it('should return true when the assessment is not under six', () => {
         expect(shouldItemBeRendered(false, item)).toBeTruthy();
-      });
-    });
-  });
-
-  describe('#calculateDomainProgress()', () => {
-    it('should return 100 when all required items have rating', () => {
-      expect(calculateDomainProgress(true, validAssessment.state.domains[0])).toEqual(100);
-      expect(calculateDomainProgress(true, validAssessment.state.domains[1])).toEqual(100);
-    });
-
-    it('should ignore not required fields', () => {
-      const domain = {
-        items: [
-          {
-            under_six_id: '1',
-            required: false,
-            rating: -1,
-          },
-          {
-            under_six_id: '2',
-            required: true,
-            rating: 1,
-          },
-          {
-            under_six_id: '3',
-            required: true,
-            rating: -1,
-          },
-        ],
-      };
-      expect(calculateDomainProgress(true, domain)).toEqual(50);
-    });
-
-    describe('when assessment is under six', () => {
-      it('should ignore fields with no under_six_id', () => {
-        const domain = {
-          items: [
-            {
-              required: true,
-              rating: -1,
-            },
-            {
-              under_six_id: '2',
-              required: true,
-              rating: 1,
-            },
-            {
-              under_six_id: '3',
-              required: true,
-              rating: 2,
-            },
-            {
-              under_six_id: '4',
-              required: true,
-              rating: -1,
-            },
-          ],
-        };
-        expect(calculateDomainProgress(true, domain)).toEqual(67);
-      });
-
-      it('should return 50 when one of two items has a rating', () => {
-        const domain = {
-          items: [
-            {
-              under_six_id: '1',
-              required: true,
-              rating: -1,
-            },
-            {
-              under_six_id: '2',
-              required: true,
-              rating: 1,
-            },
-          ],
-        };
-        expect(calculateDomainProgress(true, domain)).toEqual(50);
-      });
-    });
-
-    describe('when assessment is not under six', () => {
-      it('should ignore fields with no above_six_id', () => {
-        const domain = {
-          items: [
-            {
-              required: true,
-              rating: -1,
-            },
-            {
-              above_six_id: '2',
-              required: true,
-              rating: 1,
-            },
-            {
-              above_six_id: '3',
-              required: true,
-              rating: -1,
-            },
-          ],
-        };
-        expect(calculateDomainProgress(false, domain)).toEqual(50);
       });
     });
   });
