+++ app/javascript/Application/components/Assessment/DomainProgressBar.js
@@ -1,12 +1,18 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { calculateDomainProgress } from './AssessmentHelper';
+import { shouldItemBeRendered } from './AssessmentHelper';
 
 import './style.sass';
 
-const buildState = (isAssessmentUnderSix, domain) => ({
-  progress: calculateDomainProgress(isAssessmentUnderSix, domain),
-});
+const buildState = (isAssessmentUnderSix, domain) => {
+  const total = domain.items.filter(item => shouldItemBeRendered(isAssessmentUnderSix, item) && item.required);
+  const filled = total.filter(item => item.rating !== -1);
+  return {
+    total: total.length,
+    filled: filled.length,
+    progress: Math.round((filled.length / total.length) * 100),
+  };
+};
 
 class DomainProgressBar extends React.PureComponent {
   constructor(props) {
@@ -21,11 +27,11 @@ class DomainProgressBar extends React.PureComponent {
   }
 
   render() {
-    const { progress } = this.state;
+    const { filled, total, progress } = this.state;
     return (
       <div className={'domain-progress-bar'}>
         <div className={'progress-line'} style={{ width: `${progress}%` }} />
-        <h5 className={'progress-value'}>{`${progress}%`}</h5>
+        <h5 className={'progress-value'}>{`${filled}/${total}`}</h5>
       </div>
     );
   }
